refactor: Refactor components dir structure
