+++ components/Header/Header.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import GlobalThemeProvider from '@styles/GlobalThemeProvider';
 import theme from '@styles/theme';
-import Header from './index';
+import Header from '.';
 
 export default {
   title: 'Components/Header',
