+++ components/PostCard/PostCard.styled.ts
