+++ components/SearchBar/SearchBar.styled.ts
