+++ components/SearchBar/index.tsx
