+++ components/Typography/index.tsx
@@ -1,9 +1,1 @@
-import { StyledTypo, STypoProps } from './style';
-
-export interface TypoProps extends STypoProps {
-  children: (string | number)[] | string | number;
-}
-
-const Typography = ({ children, ...props }: TypoProps) => <StyledTypo {...props}>{children}</StyledTypo>;
-
-export default Typography;
+export { default } from './Typography';
